pyproject.toml
[project]
name = "pot-bot"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "nonebot2[fastapi]>=2.3.3",
    "nonebot-adapter-onebot>=2.4.4",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-status>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pot"]
